+++ swann-herrera/src/main/java/esgi/al/cc1/domain/User.java
@@ -13,7 +13,6 @@ public class User {
   private final int age;
 
   private User(UUID id, String firstName, String lastName, String email, int age) {
-
     this.id = id;
     this.firstName = firstName;
     this.lastName = lastName;
@@ -23,8 +22,8 @@ public class User {
 
   public static User of(UUID id, String firstName, String lastName, String email, int age)
       throws IllegalArgumentException {
-    User user = new User(id, firstName, lastName, email, age);
-    UserValidatorEngine validationEngine = UserValidatorEngine.getInstance();
+    var user = new User(id, firstName, lastName, email, age);
+    var validationEngine = UserValidatorEngine.getInstance();
     if (validationEngine.test(user)) {
       return user;
     }