+++ swann-herrera/src/main/java/esgi/al/cc1/domain/UserRepository.java
@@ -3,10 +3,13 @@ package esgi.al.cc1.domain;
 import java.util.Optional;
 import java.util.UUID;
 
+import esgi.al.cc1.kernel.Repository;
+
+@Repository
 public interface UserRepository {
   Optional<User> findById(UUID id);
 
   Optional<User> findByEmail(String email);
 
-  void add(User user);
+  void add(User user) throws IllegalArgumentException;
 }