+++ swann-herrera/src/main/java/esgi/al/cc1/domain/exception/InvalidEmail.java
@@ -1,7 +1,0 @@
-package esgi.al.cc1.domain.exception;
-
-public final class InvalidEmail extends RuntimeException {
-  public InvalidEmail(String email) {
-    super("The email " + email + " is not valid");
-  }
-}